requirements-dev.txt
# Code style and organization tools
black==23.3.0
flake8==6.0.0
flake8-docstrings==1.7.0
isort==5.12.0
mypy==1.3.0
pre-commit==3.3.2
types-requests
types-tabulate

# Type stubs for better type checking
django-stubs==4.2.7
djangorestframework-stubs==3.14.5
types-redis==4.6.0.5
types-beautifulsoup4==4.12.0.6

# Other development tools
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0

# Include production dependencies
-r requirements.txt 